chg: dev: #164: fix post-merge build issues & refactoring
